requirements.txt
asgiref==3.3.1
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.7
django-cors-headers==3.7.0
django-debug-toolbar==3.2
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
isort==5.7.0
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.5.2
Markdown==3.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
packaging==20.9
Pillow==8.1.2
psycopg2-binary==2.8.6
PyJWT==2.0.1
pylint==2.7.2
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.4
wrapt==1.12.1
